+++ src/resources/js/directives.js
@@ -2866,8 +2866,8 @@
                 onlyImages : '@onlyImages'
             },
             controller : [
-            	'$scope', '$http', '$filter', '$timeout', '$q', 'Upload', 'ServiceFoldersData', 'ServiceFilesData', 'LuyaLoading', 'AdminToastService', 'ServiceFoldersDirecotryId', 
-            	function($scope, $http, $filter, $timeout, $q, Upload, ServiceFoldersData, ServiceFilesData, LuyaLoading, AdminToastService, ServiceFoldersDirecotryId) {
+            	'$scope', '$http', '$filter', '$timeout', '$q', 'Upload', 'ServiceFoldersData', 'ServiceFilesData', 'LuyaLoading', 'AdminToastService', 'ServiceFoldersDirecotryId', 'ServiceAdminTags', 
+            	function($scope, $http, $filter, $timeout, $q, Upload, ServiceFoldersData, ServiceFilesData, LuyaLoading, AdminToastService, ServiceFoldersDirecotryId, ServiceAdminTags) {
 
                 // ServiceFoldersData inheritance
 
@@ -2880,6 +2880,14 @@
                 $scope.foldersDataReload = function() {
                     return ServiceFoldersData.load(true);
                 };
+
+                // Service Tags
+
+                $scope.tags = [];
+
+                ServiceAdminTags.load().then(function(response) {
+                    $scope.tags = response;
+                });
 
                 // ServiceFilesData inheritance
 
@@ -3240,6 +3248,7 @@
 
                 $scope.fileDetailFolder = false;
                 
+                
                 $scope.openFileDetail = function(file, force) {
                 	if ($scope.fileDetail.id == file.id && force !== true) {
                 		$scope.closeFileDetail();
@@ -3254,6 +3263,23 @@
                 		
                 		$scope.fileDetail = file;
                 	}
+                };
+
+                $scope.saveTagRelation = function(tag, file) {
+                    $http.post('admin/api-admin-storage/toggle-file-tag', {tagId: tag.id, fileId: file.id}).then(function(response) {
+                        $scope.fileDetailFull.tags = response.data;
+                    });
+                };
+
+                $scope.fileHasTag = function(tag) {
+                    var exists = false;
+                    angular.forEach($scope.fileDetailFull.tags, function(value) {
+                        if (value.id == tag.id) {
+                            exists = true;
+                        }
+                    });
+
+                    return exists;
                 };
                 
                 $scope.updateFileData = function() {